refactor: split main classes
